+++ src/parsers/contest/CodeforcesContestParser.ts
@@ -19,7 +19,7 @@ export class CodeforcesContestParser extends ContestParser {
 
   public getRegularExpressions(): RegExp[] {
     return [
-      /^https?:\/\/codeforces[.]com\/(group\/[a-zA-Z0-9]+\/)?(contest|gym)\/(\d+)$/,
+      /^https?:\/\/codeforces[.]com\/(group\/[a-zA-Z0-9]+\/)?(contest|gym)\/(\d+)(\?.*)?$/,
     ];
   }
 }